% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapStdError.R
\name{bootstrapStdError}
\alias{bootstrapStdError}
\title{Computes the standard error of estimated risk measures by the bootstrap method.}
\usage{
bootstrapStdError(dat_xts)
}
\arguments{
\item{dat_xts}{xts object containing the returns data for multiple assets with unequal returns history.}
}
\value{
The function returns the standard error of estimated risk measures by the bootstrap method. 
\item{boot_std_error}{A matrix containing standard error of estimated risk measures for different assets.}
}
\description{
This function computes the standard error of estimated risk measures by the bootstrap method.
}
\author{
Pushpak Sarkar
}
