% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructCBFData.R
\name{constructCBFData}
\alias{constructCBFData}
\title{Implements Combined Backfill for Multiple Asset Groups}
\usage{
constructCBFData(dat_xts)
}
\arguments{
\item{dat_xts}{xts object containing the returns data for multiple assets with unequal return history.}
}
\value{
This function returns the combined backfilled (CBF) dataset.
}
\description{
This function implements Combined Backfill for Multiple Asset
  Groups
}
\references{
Jiang, Y. and Martin, R. D. (2016). "Turning Long and Short Return Histories into Equal Histories: A Better Way to Backfill Returns", https://ssrn.com/abstract=2833057.
}
\author{
Pushpak Sarkar
}
