% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructGMVPortfolio.R
\name{constructGMVPortfolio}
\alias{constructGMVPortfolio}
\title{Constructs Global Minimum Variance (GMV) Portfolio.}
\usage{
constructGMVPortfolio(ret_dat)
}
\arguments{
\item{ret_dat}{xts object containing the returns data for multiple assets.}
}
\value{
The function returns the GMV portfolio weights. 
\item{w_gmv}{A vector of weights of different assets in the Global
Minimum Variance portfolio.}
}
\description{
This function Constructs Global Minimum Variance (GMV) Portfolio.
}
\author{
Pushpak Sarkar
}
