% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructPageData.R
\name{constructPageData}
\alias{constructPageData}
\title{Constructs the new dataset by bootstrapping the residuals following Page(2013)}
\usage{
constructPageData(fitted.xts, resid.mat, miss.hist.var, miss.itr, na_count,
  sReps)
}
\arguments{
\item{fitted.xts}{is the fitted data after regression.}

\item{resid.mat}{is the matrix containg residuals from linear regression.}

\item{miss.hist.var}{is the vector contating the names of the missing 
history assets.}

\item{miss.itr}{is the sorted vector containg the count of missing observations 
from the smallest to the largest.}

\item{na_count}{is the matrix containg the count of missing observations.}

\item{sReps}{is a logical flag, to indicate whether the risk measures
need to be computed for all the replicates.}
}
\value{
\item{risk.metrics}{A list containing a matrix of risk measures for each replicate.}
\item{risk.vals}{A matrix of risk measures for the whole combined backfilled dataset.}
}
\description{
Constructs the new dataset by bootstrapping the residuals following Page(2013)
}
\author{
Pushpak Sarkar
}
