% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRiskStats.R
\name{constructRiskStats}
\alias{constructRiskStats}
\title{Computes different risk measures for multiple asset returns.}
\usage{
constructRiskStats(block_dat)
}
\arguments{
\item{block_dat}{contains the returns data for multiple assets.}
}
\value{
The function returns the matrix containging risk measures. 
\item{risk_mat}{A matrix containing risk measures for different assets.}
}
\description{
This function computes different risk measures for multiple assets.
}
\author{
Pushpak Sarkar
}
