% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitValue.R
\name{fitValue}
\alias{fitValue}
\title{Implements the regression of long history asset on short history asset.}
\usage{
fitValue(short.hist.var, long.hist.var, ret.dat, num.miss)
}
\arguments{
\item{short.hist.var}{is the asset name with missing history.}

\item{long.hist.var}{is the asset name with full history.}

\item{ret.dat}{is the asset returns data containg the full history asset returns
and one or more of the short history asset returns.}

\item{num.miss}{is the number of missing observations.}
}
\value{
\item{reg.dat}{A list containing a fitted values, residuals etc.}
}
\description{
This function implements the regression of long history asset on short history asset.
}
\author{
Pushpak Sarkar
}
