% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hfunds5_ue_ts}
\alias{hfunds5_ue_ts}
\title{Monthly retrun data of five different assets.}
\format{An xts object with 123 rows and 6 columns:
\describe{
  \item{E.M.}{The asset with full history from Jan 1993 to Mar 2003}
  \item{EQUITY}{This asset has data from Jan 1994 to Mar 2003}
  \item{EVENT.1}{This asset has data from Jan 1994 to Mar 2003}
  \item{EVENT.2}{This asset has data from Jan 1994 to Mar 2003}
  \item{HEALTH}{This asset has data from Jan 1997 to Mar 2003}
  \item{HIGH.YIELD}{This asset has data from Jan 1997 to Mar 2003}
}}
\source{
TBA
}
\usage{
data("hfunds5_ue_ts")
}
\description{
A dataset containing the monthly retrun data of five different assets.
}
\keyword{data}
