% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneqhistMI.R
\name{uneqhistMI}
\alias{uneqhistMI}
\title{Implements Multiple Imputation Method folowing Page(2013)}
\usage{
uneqhistMI(dat_xts, FUN, M = 100, saveReps = FALSE)
}
\arguments{
\item{dat_xts}{xts object containing the returns data for multiple assets with unequal return history.}

\item{FUN}{indicates whether risk measures or the Global Minimum Variance (GMV) 
portfolio statistics such as portfolio weights, portfolio reurn, portfolio standard
deviation, portfolio sharpe ratio need to be computed. Two possible values
are "riskMeasures" and "gmvPortfolio".}

\item{M}{is the number of replicates for multiple imputation. Default is 100.}

\item{saveReps}{is a logical flag, to indicate whether the risk measures
and GMV portfolio statistics need to be computed for all the replicates. Default is FALSE.}
}
\value{
Based on the value of the \code{FUN} argument the function either returns the risk 
measures or the GMV portfolio statistics. If \code{saveReps} = TRUE, risk measures or 
GMV portfolio statistics are returned for each replicate else the average over all
the replicates are returned. 
\item{risk_metrics}{If \code{FUN} = "riskMeasures" and \code{saveReps} = TRUE, a list 
 of length \code{M} containing risk measures for each replicate are returned.}
\item{risk_vals}{If \code{FUN} = "riskMeasures" and \code{saveReps} = FALSE, a matrix 
 containing average of the risk measures over all replicates are returned.}
\item{gmvPortfolio_list}{If \code{FUN} = "gmvPortfolio" and \code{saveReps} = TRUE, a list 
 of length \code{M} containing GMV portfolio statistics for each replicate are retruned.}
\item{gmvPortfolio}{If \code{FUN} = "gmvPortfolio" and \code{saveReps} = FALSE, a list 
 containing average of the GMV portfolio statistics over all replicates are returned.}
}
\description{
This function implements Multiple Imputation Method folowing 
Page(2013)
}
\references{
Page, Sebastien (2013). "How to Combine Long and Short Return Histories Efficiently", Financial Analysts Journal, pp. 45-52.
}
\author{
Pushpak Sarkar
}
