% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructCBFData.R
\name{constructCBFData}
\alias{constructCBFData}
\title{Constructs the new dataset by Combined Backfill Method.}
\usage{
constructCBFData(reg.dat, new.dat, full.length)
}
\arguments{
\item{reg.dat}{is the regression data i.e. fitted data, residuals from regression.}

\item{new.dat}{is the backfilled dataset constructed for the previous 
missing history asset.}

\item{full.length}{is the length of the full history asset.}
}
\value{
\item{new.dat}{is the backfilled dataset constructed by CBF.}
}
\description{
Constructs the new dataset by Combined Backfill Method.
}
\author{
Pushpak Sarkar
}
