% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneqhistCBF.R, R/uneqhistPage.R
\name{fitValue}
\alias{fitValue}
\alias{fitValue}
\title{Implements Combined Backfill for Multiple Asset Groups}
\usage{
fitValue(short.hist.var, long.hist.var, ret.dat, num.miss)

fitValue(short.hist.var, long.hist.var, ret.dat, num.miss)
}
\arguments{
\item{dat.mat}{is the returns data for multiple assets with unequal return history.}

\item{saveReps}{is a logical flag, to indicate whether the risk measures
need to be computed for all the replicates. Default is FALSE.}

\item{dat.mat}{is the returns data for multiple assets with unequal return history.}

\item{saveReps}{is a logical flag, to indicate whether the risk measures
need to be computed for all the replicates. Default is FALSE.}
}
\value{
\item{risk.metrics}{A list containing a matrix of risk measures for each replicate.}
\item{risk.vals}{A matrix of risk measures for the whole combined backfilled dataset.}

\item{risk.list}{A list or matrix containing risk measures for either the whole dataset or for each replicate.}
}
\description{
This function implements Combined Backfill for Multiple Asset Groups

This function implements Multiple Imputation Method folowing 
Page(2013)
}
\author{
Pushpak Sarkar

Pushpak Sarkar
}
