% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneqhistCBF.R
\name{uneqhistCBF}
\alias{uneqhistCBF}
\title{Implements Combined Backfill for Multiple Asset Groups}
\usage{
uneqhistCBF(dat_mat, gmvPortfolio = FALSE)
}
\arguments{
\item{dat_mat}{is the returns data for multiple assets with unequal return history.}

\item{gmvPortfolio}{is a logical flag, to indicate whether the Global Minimum Variance (GMV) 
portfolio such as portfolio weights, portfolio reurn, portfolio standard
deviation, portfolio sharpe ratio need to be computed. Default is FALSE.}
}
\value{
\item{risk_metrics}{A matrix of risk measures for the whole combined backfilled dataset.}
\item{output}{A list containing the matrix of risk measures for the whole combined backfilled dataset and the Global Minimum Variance portfolio weights, portfolio return, portfolio standard deviation, portfolio sharpe ratio}
}
\description{
This function implements Combined Backfill for Multiple Asset Groups
}
\author{
Pushpak Sarkar
}
