% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneqhistCBF.R
\name{uneqhistCBF}
\alias{uneqhistCBF}
\title{Implements Combined Backfill for Multiple Asset Groups}
\usage{
uneqhistCBF(dat.mat, saveReps = FALSE)
}
\arguments{
\item{dat.mat}{is the returns data for multiple assets with unequal return history.}

\item{saveReps}{is a logical flag, to indicate whether the risk measures
need to be computed for all the replicates. Default is FALSE.}
}
\value{
\item{risk.metrics}{A list containing a matrix of risk measures for each replicate.}
\item{risk.vals}{A matrix of risk measures for the whole combined backfilled dataset.}
}
\description{
This function implements Combined Backfill for Multiple Asset Groups
}
\author{
Pushpak Sarkar
}
