% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneqhistMI.R
\name{uneqhistMI}
\alias{uneqhistMI}
\title{Implements Multiple Imputation Method folowing Page(2013)}
\usage{
uneqhistMI(dat.mat, saveReps = FALSE)
}
\arguments{
\item{dat.mat}{is the returns data for multiple assets with unequal return history.}

\item{saveReps}{is a logical flag, to indicate whether the risk measures
need to be computed for all the replicates. Default is FALSE.}
}
\value{
\item{risk.list}{A list or matrix containing risk measures for either the whole dataset or for each replicate.}
}
\description{
This function implements Multiple Imputation Method folowing 
Page(2013)
}
\author{
Pushpak Sarkar
}
